requirements/requirements.txt
alembic==1.13.1
fastapi==0.115.12
Pillow==11.2.1
pydantic==2.11.3
python-dotenv==1.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.35
torchvision==0.20.1
ultralytics==8.3.58
uvicorn==0.34.2
